requirements.txt
# Core Dependencies
faster-whisper==1.1.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12

# Video Processing
ffmpeg-python>=0.2.0

# Database & Cache (optional - can use SQLite)
aiosqlite>=0.20.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=24.1.0

# Monitoring & Logging
loguru>=0.7.2

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.6.0

